tests/test_roman_numeral.py
"""Tests for the RomanNumeral class."""
import pytest
from models.roman_numeral import RomanNumeral
from models.scale_info import ScaleInfo
from models.note import Note

def test_basic_roman_numerals() -> None:
    """Test basic Roman numeral parsing."""
    scale_info = ScaleInfo(root=Note.from_str("C"), mode="major")
    
    # Test major scale numerals
    assert RomanNumeral.from_str("I", scale_info).scale_degree == 1
    assert RomanNumeral.from_str("IV", scale_info).scale_degree == 4
    assert RomanNumeral.from_str("V", scale_info).scale_degree == 5
    
    # Test minor scale numerals
    assert RomanNumeral.from_str("i", scale_info).scale_degree == 1
    assert RomanNumeral.from_str("iv", scale_info).scale_degree == 4
    assert RomanNumeral.from_str("v", scale_info).scale_degree == 5

def test_chord_qualities() -> None:
    """Test chord quality determination."""
    major_scale = ScaleInfo(root=Note.from_str("C"), mode="major")
    minor_scale = ScaleInfo(root=Note.from_str("C"), mode="minor")
    
    # Test major scale qualities
    assert RomanNumeral.from_str("I", major_scale).chord_quality == "major"
    assert RomanNumeral.from_str("ii", major_scale).chord_quality == "minor"
    assert RomanNumeral.from_str("vii°", major_scale).chord_quality == "diminished"
    
    # Test minor scale qualities
    assert RomanNumeral.from_str("i", minor_scale).chord_quality == "minor"
    assert RomanNumeral.from_str("III", minor_scale).chord_quality == "major"
    assert RomanNumeral.from_str("ii°", minor_scale).chord_quality == "diminished"

def test_flattened_numerals() -> None:
    """Test flattened Roman numerals."""
    scale_info = ScaleInfo(root=Note.from_str("C"), mode="major")
    
    rn = RomanNumeral.from_str("bIII", scale_info)
    assert rn.is_flattened
    assert rn.scale_degree == 3
    assert rn.get_note() == Note.from_str("Eb")
    
    rn = RomanNumeral.from_str("bVI", scale_info)
    assert rn.is_flattened
    assert rn.scale_degree == 6
    assert rn.get_note() == Note.from_str("Ab")

def test_chord_modifiers() -> None:
    """Test chord modifiers."""
    scale_info = ScaleInfo(root=Note.from_str("C"), mode="major")
    
    # Test seventh chords
    assert RomanNumeral.from_str("V7", scale_info).chord_quality == "dominant7"
    assert RomanNumeral.from_str("Imaj7", scale_info).chord_quality == "major7"
    assert RomanNumeral.from_str("ii7", scale_info).chord_quality == "minor7"
    
    # Test other modifiers
    assert RomanNumeral.from_str("vii°7", scale_info).chord_quality == "diminished7"
    assert RomanNumeral.from_str("iiø7", scale_info).chord_quality == "half-diminished7"
    assert RomanNumeral.from_str("V+", scale_info).chord_quality == "augmented"


# Repeat for other test functions

def test_note_generation() -> None:
    """Test note generation from Roman numerals."""
    c_major = ScaleInfo(root=Note.from_str("C"), mode="major")
    c_minor = ScaleInfo(root=Note.from_str("C"), mode="minor")
    
    # Test in C major
    assert RomanNumeral.from_str("I", c_major).get_note() == Note.from_str("C")
    assert RomanNumeral.from_str("IV", c_major).get_note() == Note.from_str("F")
    assert RomanNumeral.from_str("V", c_major).get_note() == Note.from_str("G")
    
    # Test in C minor
    assert RomanNumeral.from_str("i", c_minor).get_note() == Note.from_str("C")
    assert RomanNumeral.from_str("bIII", c_minor).get_note() == Note.from_str("Eb")
    assert RomanNumeral.from_str("v", c_minor).get_note() == Note.from_str("G")

def test_invalid_numerals() -> None:
    """Test invalid Roman numerals."""
    scale_info = ScaleInfo(root=Note.from_str("C"), mode="major")
    
    with pytest.raises(ValueError):
        RomanNumeral.from_str("VIII", scale_info)  # Invalid numeral
    
    with pytest.raises(ValueError):
        RomanNumeral.from_str("X", scale_info)  # Invalid numeral

def test_roman_numeral_creation() -> None:
    # Test implementation here
    pass

def test_roman_numeral_quality() -> None:
    # Test implementation here
    pass
